feat(test): Assert invalid config raises ValueError
This patch introduces a test case that illustrates the inability of the
configuration validation implementation to detect invalid entries for
nested keys. This is illustrated by overwriting the `fiets of E-fiets`
entry for a valid starting configuration by either a list containing an
invalid value or a scalar value of this invalid value.

The failing test is resolved in following commits.