fix: Do attempt to fix `geavanceerd` when present
If the `geavanceerd` key is present, it might still be possible to
resolve part of the keys that might have been present within the
outdated `verdeling` key.

This patch removes the early return based on the presence of the
`geavanceerd` key and continues with the fix attempt. Only when the
deprecated key itself (`verdeling`) is not present at all, we exit
early.