feat: Add fixer for `fiets of E-fiets` entries
The previous commit has updated the `fiets of E-fiets` entry from being
a checklist to a checkbox, resulting in different entries for the config
files.

This adds a fixer routine that transforms the old lists, like
`['Fiets']` into the matching new configuration format in an attempt to
automatically resolve old configuration files.