fix: Add conditional use citizens/places_of_employment
The generic implementation in `competition.py` should either operate on
`citizens_per_class` or `places_of_employment` and the used data is
selected conditionally through the boolean flag `citizens` and assigned
into `citizens_or_places_of_employment`.

It turned out that towards the end of the script a plain instance of
`citizens_per_class` per class was used, without the conditional
selection based on the value of `citizens`.

This patch makes sure the conditional values in
`citizens_or_places_of_employment` are used here too.

The tests are updated accordingly.