fix: Use `citizens_per_class` or `places_of_employment` for income distributions
The initialization of `income_distributions` only considered the
`citizens_per_class` input data and did not contain a switch to use the
`places_of_employment` whenever the flag `citizens` is False.

That inconsistency is addressed in this patch by moving the
initialization of the `income_distributions` array into a separate
function to which we either supply the `citizens_per_class` or
`places_of_employment` data depending on the context.